requirements.txt
streamlit
pandas
python-dotenv
jira
requests
plotly
